index: use a hashmap to speed up path-based lookups
We still need to insert the entries into a sorted vector due to the
sorted nature of the index, so insertion should still be just as
slow. But lookups can now happen in O(1). There are still a few places
where we could make more use of the hashmap to search for collisions but
this commit represents a baseline.